Main/Pause Menus and Camera Scalar
I added a main menu scene, a pause menu to the existing "game" scene, and the ability to change the sensitivity of the camera with a slider. There are a couple bugs to iron out but I am out of time before the meeting. There is also a lot of work to be done because the features are bare bones.